Update Count Inversions
Problem link: https://www.geeksforgeeks.org/problems/inversion-of-array-1587115620/1

// Problem Statement : Given an array of integers. Find the Inversion Count in the array. 
Inversion Count: For an array, inversion count indicates how far (or close) the array is from being sorted. If the array is already sorted then the inversion count is 0.
If an array is sorted in the reverse order then the inversion count is the maximum.